requirements.txt
streamlit
langchain
langchain-groq
groq
python-dotenv
pygments
streamlit-ace